+++ zellij-server/src/screen.rs
@@ -758,6 +758,7 @@ pub enum ScreenInstruction {
         PaneId,
         Option<String>,
         Option<String>,
+        Option<String>,
         Option<NotificationEnd>,
     ),
     WriteKeyToPaneId(
@@ -9400,11 +9401,11 @@ pub(crate) fn screen_thread_main(
                 }
                 screen.render(None)?;
             },
-            ScreenInstruction::SetPaneColor(pane_id, fg, bg, _completion) => {
+            ScreenInstruction::SetPaneColor(pane_id, fg, bg, frame, _completion) => {
                 let all_tabs = screen.get_tabs_mut();
                 for tab in all_tabs.values_mut() {
                     if tab.has_pane_with_pid(&pane_id) {
-                        tab.set_pane_color(pane_id, fg, bg).non_fatal();
+                        tab.set_pane_color(pane_id, fg, bg, frame).non_fatal();
                         break;
                     }
                 }